moving logging.go into its own internal library
